# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /microfilm-api
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /microfilms:
    # our controller name
    x-swagger-router-controller: microfilm
    get:
      operationId: getAll
      description: get the microfilm list
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Microfilm"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: save
      description: add a new roll to the list
      # movie info to be stored
      parameters:
        - name: roll_no
          description: Movie properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Microfilm"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /newspapers:
    # our controller name
    x-swagger-router-controller: microfilm
    get:
      operationId: getAllNewspapers
      description: get the newspapers list
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Newspaper"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /search:
    x-swagger-router-controller: microfilm
    get: 
      operationId: searchMicrofilms
      description: search Microfilms
      parameters: 
        - $ref: "#/parameters/newspaper"
        - $ref: "#/parameters/date_start"
        - $ref: "#/parameters/date_end"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Microfilm"
        default: 
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"


  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  GetMicrofilmsListResponse:
    required:
      - microfilms
    properties:
    # The array of microfilms
      microfilms:
        type: array
        items:
          type: object
          properties:
            roll_no:
              type: string
            newspaper:
              type: string
            date_start:
              type: string
              format: date
            date_end:
              type: string
              format: date
              
  Newspaper:
    properties:
      name: 
        description: name of the newspaper
        type: string


  Microfilms:
    type: array
    items:
      $ref: "#/definitions/Microfilm"
  
  Microfilm:
     properties:
      roll_no:
        type: string
        description: task object name
      newspaper:
        type: string
        description: task description
      date_start:
        type: string
        format: date
        description: task object name
      date_end:
        type: string
        format: date
        description: task description        

  GetRollNoResponse:
    required:
      - roll_no
      - newspaper
      - date_start
      - date_end
    properties:
      roll_no:
        type: string
        description: task object name
      newspaper:
        type: string
        description: task description
      date_start:
        type: string
        format: date
        description: task object name
      date_end:
        type: string
        format: date
        description: task description

  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
parameters:
  newspaper:
    name: newspaper
    in: query
    description: Newspaper name
    type: string
    required: false 
  date_start:
    name: date_start
    in: query
    description: Search starts from this date
    type: string
    format: date
    required: true
  date_end:
    name: date_end
    in: query
    description: Search for until this date
    type: string
    format: date
    required: true